[CAZ-2180] Create endpoint to get Account's Mandates for requested CleanAirZone
