+++ src/it/resources/data/json/response/direct-debit-mandates-for-caz-response.json
@@ -1,0 +1,12 @@
+{
+  "mandates": [
+    {
+      "id": "i4nuo03jginfke5c3ebrvgig6a",
+      "status": "CREATED"
+    },
+    {
+      "id": "j4nuo03jginfke5c3ebrvgig6a",
+      "status": "FAILED"
+    }
+  ]
+}
