+++ src/main/java/uk/gov/caz/psr/controller/DirectDebitMandatesController.java
@@ -9,10 +9,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.caz.psr.dto.CreateDirectDebitMandateRequest;
 import uk.gov.caz.psr.dto.CreateDirectDebitMandateResponse;
+import uk.gov.caz.psr.dto.directdebit.DirectDebitMandatesForCazResponse;
 import uk.gov.caz.psr.dto.directdebit.DirectDebitMandatesResponse;
 import uk.gov.caz.psr.model.directdebit.CleanAirZoneWithMandates;
+import uk.gov.caz.psr.model.directdebit.Mandate;
 import uk.gov.caz.psr.service.directdebit.DirectDebitMandatesService;
 import uk.gov.caz.psr.util.CleanAirZoneWithMandatesToDtoConverter;
+import uk.gov.caz.psr.util.MandatesToDtoConverter;
 
 @RestController
 @AllArgsConstructor
@@ -20,15 +23,26 @@ import uk.gov.caz.psr.util.CleanAirZoneWithMandatesToDtoConverter;
 public class DirectDebitMandatesController implements DirectDebitMandatesControllerApiSpec {
 
   public static final String BASE_PATH = "/v1/payments/accounts/{accountId}/direct-debit-mandates";
+  public static final String FOR_CAZ_PATH = BASE_PATH + "/{cleanAirZoneId}";
 
   private final DirectDebitMandatesService directDebitMandatesService;
   private final CleanAirZoneWithMandatesToDtoConverter converter;
+  private final MandatesToDtoConverter mandatesToDtoConverter;
 
   @Override
   public ResponseEntity<DirectDebitMandatesResponse> getDirectDebitMandates(UUID accountId) {
     List<CleanAirZoneWithMandates> directDebitMandates = directDebitMandatesService
         .getDirectDebitMandates(accountId);
     return ResponseEntity.ok(converter.toDirectDebitMandatesResponse(directDebitMandates));
+  }
+
+  @Override
+  public ResponseEntity<DirectDebitMandatesForCazResponse> getDirectDebitMandatesForCaz(
+      UUID accountId, UUID cleanAirZoneId) {
+    List<Mandate> mandates = directDebitMandatesService
+        .getMandatesForCazWithStatus(accountId, cleanAirZoneId);
+    return ResponseEntity
+        .ok(mandatesToDtoConverter.toDirectDebitMandatesResponse(mandates));
   }
 
   @Override
