+++ src/main/java/uk/gov/caz/psr/controller/DirectDebitMandatesControllerApiSpec.java
@@ -17,20 +17,20 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import uk.gov.caz.psr.dto.CreateDirectDebitMandateRequest;
 import uk.gov.caz.psr.dto.CreateDirectDebitMandateResponse;
+import uk.gov.caz.psr.dto.directdebit.DirectDebitMandatesForCazResponse;
 import uk.gov.caz.psr.dto.directdebit.DirectDebitMandatesResponse;
 
 /**
  * Direct Debit mandates REST controller API specification.
  */
-@RequestMapping(value = DirectDebitMandatesController.BASE_PATH,
-    produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE)
 public interface DirectDebitMandatesControllerApiSpec {
 
   /**
    * Gets a list of clean air zones with associates direct debit mandates for a given {@code
    * accountId}.
    */
-  @GetMapping
+  @GetMapping(DirectDebitMandatesController.BASE_PATH)
   @ApiOperation(value = "${swagger.operations.payments.direct-debit-mandates.description}",
       response = DirectDebitMandatesResponse.class)
   @ApiResponses({
@@ -48,9 +48,29 @@ public interface DirectDebitMandatesControllerApiSpec {
       @PathVariable("accountId") UUID accountId);
 
   /**
-   * Allows Fleets Front-end to create a new DirectDebitMandate assigned to company.
-   * It creates the direct-debit-mandate in GOV.UK PAY and returns next steps which
-   * needs to be completed.
+   * Gets a list of direct debit mandates for a given {@code accountId} and {@code cleanAirZoneId}.
+   */
+  @GetMapping(DirectDebitMandatesController.FOR_CAZ_PATH)
+  @ApiOperation(value = "${swagger.operations.payments.direct-debit-mandates-for-caz.description}",
+      response = DirectDebitMandatesForCazResponse.class)
+  @ApiResponses({
+      @ApiResponse(code = 500,
+          message = "Internal Server Error / No message available"),
+      @ApiResponse(code = 400, message = "Correlation Id missing"),
+      @ApiResponse(code = 200, message = "Direct debit mandates for request clean air zone"),})
+  @ApiImplicitParams({@ApiImplicitParam(name = "X-Correlation-ID",
+      required = true,
+      value = "CorrelationID to track the request from the API gateway through"
+          + " the Enquiries stack",
+      paramType = "header")})
+  @ResponseStatus(HttpStatus.OK)
+  ResponseEntity<DirectDebitMandatesForCazResponse> getDirectDebitMandatesForCaz(
+      @PathVariable("accountId") UUID accountId,
+      @PathVariable("cleanAirZoneId") UUID cleanAirZoneId);
+
+  /**
+   * Allows Fleets Front-end to create a new DirectDebitMandate assigned to company. It creates the
+   * direct-debit-mandate in GOV.UK PAY and returns next steps which needs to be completed.
    *
    * @return {@link CreateDirectDebitMandateResponse} wrapped in {@link ResponseEntity}.
    */
@@ -65,7 +85,8 @@ public interface DirectDebitMandatesControllerApiSpec {
       required = true,
       value = "UUID formatted string to track the request through the enquiries stack",
       paramType = "header")})
-  @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
+  @PostMapping(value = DirectDebitMandatesController.BASE_PATH,
+      consumes = MediaType.APPLICATION_JSON_VALUE)
   @ResponseStatus(HttpStatus.CREATED)
   ResponseEntity<CreateDirectDebitMandateResponse> createDirectDebitMandate(
       @PathVariable("accountId") UUID accountId,
